feat: Inter改为Long

+++ base/common/src/main/java/com/kewen/framework/base/common/model/DeptPrimary.java
@@ -14,7 +14,7 @@ public class DeptPrimary extends Dept {
     public DeptPrimary() {
     }
 
-    public DeptPrimary(Integer id, String name, Boolean isPrimary) {
+    public DeptPrimary(Long id, String name, Boolean isPrimary) {
         super(id, name);
         this.isPrimary = isPrimary;
     }
